Arreglando problema con modelo de data y agregando la posibilidad de que el demonio mqtt tome datos sin cifrar

+++ elektron_django/data/models.py
@@ -20,7 +20,7 @@ def to_UTC(date):
 
 def remove_null_data(data):
     data = data if data != None else 0
-    data = int(data)
+    data = float(data)
     return data
 
 class Data(models.Model):
